+++ languages/en.json
@@ -1,5 +1,5 @@
 {
-  "OSE-CharacterBuilder": {
+  "osr-character-builder": {
     "formAttrib": "Character Atributes",
     "formStr": "STR",
     "formInt": "INT",
