+++ styles/main.css
@@ -4,7 +4,7 @@
   --glow-b: #1b4e1b
 }
 /*  character sheet - character builder button */
-.ose-choose-class {
+.osr-choose-class {
   color: #55ff55;
   animation-delay: 1s;
   animation: flash 1.6s infinite;
